LayoutTests/imported/w3c/web-platform-tests/css/css-contain/container-queries/animation-container-size-expected.txt

Test

FAIL Animation affects container query evaluation assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 255)"
